+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Change
 
 - don't send quality change sample if the quality did not change
+- Bitmovin, ExoPlayer: changed player error mapping to improve transparency (AN-1507)
 - set DRM information for each event data sent to server
 
 ## v1.15.0